# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"depth-chart-controller","description":"Depth Chart Controller"}],"paths":{"/players":{"get":{"tags":["depth-chart-controller"],"summary":"getPlayers","operationId":"getPlayersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Player"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/players/{playerName}":{"put":{"tags":["depth-chart-controller"],"summary":"addPlayer","operationId":"addPlayerUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"playerName","in":"path","description":"playerName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Player"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/sports/{sportId}":{"get":{"tags":["depth-chart-controller"],"summary":"getFullDepthChart","operationId":"getFullDepthChartUsingGET","produces":["*/*"],"parameters":[{"name":"sportId","in":"path","description":"sportId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DepthChartDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/sports/{sportId}/player/{playerId}/position/{positionId}":{"get":{"tags":["depth-chart-controller"],"summary":"getPlayersUnderPlayerInDepthChart","operationId":"getPlayersUnderPlayerInDepthChartUsingGET","produces":["*/*"],"parameters":[{"name":"playerId","in":"path","description":"playerId","required":true,"type":"integer","format":"int64"},{"name":"positionId","in":"path","description":"positionId","required":true,"type":"string"},{"name":"sportId","in":"path","description":"sportId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DepthChartDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["depth-chart-controller"],"summary":"removePlayerFromDepthChart","operationId":"removePlayerFromDepthChartUsingDELETE","produces":["*/*"],"parameters":[{"name":"playerId","in":"path","description":"playerId","required":true,"type":"integer","format":"int64"},{"name":"positionId","in":"path","description":"positionId","required":true,"type":"string"},{"name":"sportId","in":"path","description":"sportId","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/sports/{sportId}/player/{playerId}/position/{positionId}/{depth}":{"put":{"tags":["depth-chart-controller"],"summary":"addPlayerToDepthChart","operationId":"addPlayerToDepthChartUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"depth","in":"path","description":"depth","required":true,"type":"integer","format":"int32"},{"name":"playerId","in":"path","description":"playerId","required":true,"type":"integer","format":"int64"},{"name":"positionId","in":"path","description":"positionId","required":true,"type":"string"},{"name":"sportId","in":"path","description":"sportId","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"DepthChartDto":{"type":"object","properties":{"positionPlayersMap":{"type":"object","additionalProperties":{"type":"array","items":{"type":"integer","format":"int64"}}},"sport":{"type":"string"}},"title":"DepthChartDto"},"Player":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"Player"}}}